requirements.txt
Pillow==10.1.0
PySide6==6.6.0